[INTERNAL][FIX] sap.m.NavContainer: reduce failure of unit test
Since recently, the test "NavContainer in Popover: Focus is changed only
after transition" regularly runs into a timeout. This seems to happen
due to the following chain of events:
 - the test renders a button (oOpeningBtn) into the #qunit-fixture
 - on 'AfterRendering' of that button, the test opens a Popover, taking
   the button as openBy reference point
 - the test then waits for the 'afterOpen' event of the Popover
 - if the ResizeHandler is triggered while the Popover is opening,
   then some code in Popover.prototype._applyPosition detects that the
   reference point is now off-screen (which is true as #qunit-fixture by
   default has off-screen coordinates) and auto-closes the Popover
 - if this happens, then the test waits endlessly for the 'afterOpen'
   event

Due to the timing aspect (check interval of ResizeHandler), the issue
does not occur for each execution. It however became much more likely
recently due to change Ic969aab39189f2b70bc6a0c05bd3ab84bd8d2491 which
enabled resize events on the Popover while it is in the "opening" state.

As a workaround, this change moves the qunit-fixture into the visible
area for the duration of the above mentioned test. This prevents the
auto-close of the Popover.

Note: repeated execution of the NavContainer test still shows errors in
other test cases from time to time. Those need to be analyzed and fixed
separately.

Change-Id: If38d340ebe41b797e7cebdc3cc5ddcf73093bd34