[INTERNAL] core.rules.View.support.js: do not complain about sub views
The support assistant rule 'xmlViewWrongNamespace' checks the content of
all XMLView instances whether it uses sap.ui.core.mvc as namespace of
the root tag.

However, for "subviews" which are created behind the scenes to embed
native HTML into the UI5 control tree of an outer XMLView, this never is
true as those "subviews" always have a root node from the XHTML
namespace.

Similary, such "subviews" don't necessarily have a default namespace
(as checked by rule 'xmlViewDefaultNamespace').

Last, but not least, "subviews" are created from already known XML
content, so sync/async loading of the content does not apply for them
(as checked by rule 'asynchronousXMLViews').

With this change, all three rules are adapted to ignore "subviews".

Note that the term "subviews" does not apply to nested XMLViews that are
declared in an XMLView's source file.

Change-Id: I7b585dc67489efc412592ea4762a76c0ffced569