[INTERNAL] mdc.Table: Suppress grouping on non-visible column in ResponsiveTable
Actually the v4.ODataModel does not analyze the current sorter to enhance $select statement when autoExpandSelect: true
and the property it's not visible as a column in the table but the user still decide to group by it.
This leads to several exceptions thrown by the ODataModel and to a wrong data visualisation on the UI.

With this change we always suppress the grouping by non-visible columns.

BCP: 2280102655
CR-Id: 002075125900001415902022
Change-Id: Ie5f38bfc4cc87399a4e30fd8000395828f9867af
(cherry picked from commit 6e2e127310af52bca0114d8ed1ea4ecd2b9dba8a)