[INTERNAL][FIX]sap.ui.fl: CompVariant missing property
In change 5622618 the mapping of properties to file content for the
CompVariant was incorrect for the persistencyKey, which should be
inside a "selector" object.
Additionally, the generator property was always being directly written
on FlexObjectFactory.createCompVariant - now it can also be passed as
part of the file content.

Change-Id: I5f62dd2562797a18549e332f8e2a1d7aa1171e6f
BCP: 2280108988
(cherry picked from commit 6fd91b3d0c666338d0837dbdab0d54a54ca2f8a4)
CR-Id: 002075125900001901982022