[INTERNAL] sap.m.Select: Correct picker labels update
Background: Updating of the pickers ariaLabelledBy was resulting in flickering visual issue, if the popover was closing at the same time.
A change in behaviour was introduced and the valueState update that leads to pickers labels update was postponed and in some cases the valueState property setting
was behaving asynchronous.

Solution: Move the postponed call to the setter of the pickers ariaLabelledBy association.

BCP: 2280087874
BCP: 002075129400001188692022
Fixes: https://github.com/SAP/openui5/issues/3529
Change-Id: I336b319ee34980726c99f5cd22889332918710ea