[FIX] sap.ui.model.Binding#destroy: ensure that binding is removed from the model's active bindings list
While creating a POC for CPOUI5MODELS-763 we figured out that there is
a potential memory leak if a "change" event handler for a binding is not
detached before the binding gets destroyed. It has to be ensured that
the binding is removed from the model's "active" bindings list when the
binding is being destroyed.

PS1: done
PS2: fixed typo and added JSDoc
PS3: fixed instable integration test
PS4: Review comments (use term event handler instead of listener)

Change-Id: I3989209e905e32321663e068f776df0d6aa6ae30