[INTERNAL] ItemBaseFlex: react gracefully on certain events
In case a change appliance will be triggered and the change handling
has already been performed (for example through a deeper layer)
the according change should be marked as not applicable.

Rejecting the change handler will ensure that the according revert is
not going to be triggered, which will ensure the proper order of reverts
similar as in the apply chain.

Examples:
- adding the same item multiple times
- removing the same item multiple times
- moving an item that has already been removed

BCP: 002075129400003913842022
Change-Id: I82c784b4e94bc01bb6da046330e3025e8eba2a59