[INTERNAL] sap.m.DateTimePicker: pickers properly handled on phone
Issue: When the DTP is used on small screen and the user chooses
a date from the Calendar with Enter, the Calendar wasn't hiding.

Solution: When the user chooses a date from the Calendar, the DTP
sets a delegate to the Calendar after rendering to ensure that it
becomes hidden and the TimePicker is visualised.

Change-Id: I785d77eb47969115b360f13bca41adde1007c4b3