[FIX] Demo Kit: DataTable filter row formatting fixed
Problem: When DataTables are nested, the filtering `tr` gets erroneously appended to the nested table's header, instead of the main one. As a result, the main table's filter row renders placeholder cells with no filters inside.

Solution: Change the specifcity of the selector responsible for obtaining the `thead`, making sure only the main is selected.

BCP: 2280110637
Change-Id: I4383b4a48cf7792cf44ed2ada0cacbbafba9f626