[FIX] P13nConditionPanel now respects upperCase format on field switch
Problem: With existing lowercase value on the value field
on switching to keyfield which is annotated with display
format "UpperCase", the annotation is not taken
into consideration and the value will remain lowercase.

Solution: On switching and updating the key field we check
whether the selected field have UpperCase display format
and apply it to the value.

BCP: 002075129400002594982022
Change-Id: Iad0e362a784a1169fd7ba0944addc632b3d966ef
CR-Id: 002075125900001780052022