[FIX] sap.m.IllustratedMessage: Fixed Illustration not loading initially
Problem: Illustration isn't loaded when 'IllustratedMessageSize' is
set to different than 'Auto' initially.

Root Cause: There was a slight refactoring due to a new feature. We are
not calling the new private "_updateSymbol" where we should in the
'_updateDomSize' function when we have a custom 'IllustratedMessageSize'
set initially.

Solution: We call the '_updateSymbol' when we have
a custom 'IllustratedMessageSize' set initially.

Additionally an issue where the size of the illustration won't update
when change it from 'Auto' to a different one and back to 'Auto'
is now fixed.

BCP: 2280118931
Change-Id: I824883636f70b5b7e5723c2a4afdfdc034bb4958
(cherry picked from commit 2351e9eee34bb31f483013306cca999712f8ac22)
CR-Id: 002075125900001757732022