[FIX] sap.f.DynamicPage: Disappear of header is fixed
Problem: When navigating from one page of the App to another, and the
first one, containing DynamicPage (headerExpanded=true and
preserveHeaderStateOnScroll=true), is scrolled down, the header gets
snapped (overridePreserveHeaderStateOnScroll is called). Then, when
navigating back to the first page, the header disappears.

Solution: The problem happens during the transitioning from one
NavContainer to another. There is a point at which the whole height of
the content of the DynamicPage is 0, but the header still has height.
As, resize handlers (both onResize of the DynamicPage and
onChildControlsHeight change) are called at this moment,
overridePreserveHeaderStateOnScroll is called, and as the header seems
to be bigger than allowed (compared to the whole height of the
DynamicPage), it gets snapped.
Now, we do not override preserveHeaderStateOnScroll property and do not
snap the header, if width or height of the DynamicPage is 0 (means that
it's getting out of view).

BCP: 2270089016
Change-Id: I3536155128446efcb0587d5d4c1fcc8eb22cd46c
CR-Id: 002075125900001723112022
(cherry picked from commit ad219f8d11b40243863a5224d69f70993abdc541)