[FIX] sap.m.DynamicDateRange: Correct dates are now set to the calendar
We use bUTC flag in the creation of the controls, because we have cases
in which the formatter will be set to UTC. This will lead to a specific,
parsing and formatting of the value, but it also affect the setters in
the controls we create, hence we have to remove the offset before we
set the values to the controls, so we don't have any shifting in dates.
Providing a test was too complex for this case, as we need to have more
specific configuration and information for local/configured timezones.

BCP: 2280121072

CR-Id: 002075125900002074242022
Change-Id: Ib327ae998248da68dd8e7f15b51e6184644b6a12
(cherry picked from commit e6a2afe7e43e27255a1574dd0caa2922432e54de)