[FIX] sap.f.DynamicPage: Snapping issue corrected
Problem:
Incorrect positioning of the stickySubheader upon
expanding with a click when the scroll position
is close to the snap breakpoint.

The issue occured because of de-synchronization
between (1)snapping the header and (2) snapping
the [sticky] subHeader.
So the _shouldStickStickyContent function returned
false (instead of true), because it expected the
header to be expanded in the content area, while
the header was expanded in the title area.

One solution is to add an extra check into
the _shouldStickStickyContent to determine the
position of the header.However that is not enough to
solve all cases, because in those other cases the
position of the header is determined ^after^ the
_shouldStickStickyContent is called.

Solution:
1) add an extra check into
the _shouldStickStickyContent to determine the
position of the header
2) use the same breakpoint to toggle both the header
and the [sticky] subHeader of the DynamicPage at the same
time (since change Icfc7ed931ebefd2a630a7d5149e75f677d094322
we can do that, obstacles were removed there).

BCP: 2280092564
Change-Id: I9cf96f3f1dcc8c53aed527747109e73e9fdd3f23