[FIX] Demo Kit: DataTable filter list population fixed
Problem: When a filter of type "select" is used in a DataTable, some of the options are not included in the list.

Root cause: When `p` elements are used instead of `li` for listing filters, DataTable.prototype.getSelectOptions does not include them in aOptions.

Solution: Introduce additional handling for paragraphs.

BCP: 2280110640
Change-Id: I063e9a357d196e25154e5d76c27e304d3e80aa2b