[INTERNAL] ColumnResizer: fix auto column resize on mouse double click
- The auto size column should also consider the sort indicators. Since
the sort indicator is added via '::after' element, we now make use of
the getComputedStyle function to get the width and padding-left and
padding-right information, to add it to the column width. This will
ensure correct auto width calculation including the sort indicator.
- We also allow truncation of inner flex items for the div with class
"sapMColumnHeader". This will prevent unexpected column header overflow
and truncate as expected.
BCP: 2280094078

Change-Id: I46568a561ab67f5d40f9567bdda6d2455fd55511