[FIX] sap.uxap.AnchorBar: Corrected qunit
Problem:
Unexpected extra assertions in the qunit,
because a function was called more often than expected.

Solution:
Since the additional call is not subject to the test and is not problematic,
fixed by reseting the couter for function calls before the actual test.
Also switched to synchronous call of the resize listener to speed-up the test.

BCP: 2280125573
Change-Id: I8e5b740cfe54f981537e78743b90b8a2a48c8091
(cherry picked from commit 70e54750aa21350dc781b59ed5f6bfc5101df5f8)
CR-Id: 002075125900002001372022