[INTERNAL][FEATURE] sap.m.DynamicDateUtil: Now provides a way to remove offset caused by timezone.
We are downporting this new functionality to help the application developers
to handle the newly introduced feature with timezones. As per the incident
related here, they have tested the fix in master and it works for them, so
they ask for transport for 1.102. The newly added feature is a part of a util
and not new functionality for any of the controls, this meas that the current
behaviour will not be changed in anyway, we just provide one more tool for
the developers to fix their issues.

BCP: 2280126776
CR-Id: 002075125900002072202022
Change-Id: I8fc7078191b8c0c5258ecfb561ddd60a6ea2ba9c
(cherry picked from commit 5b956ec94192799866b121a5897766d4246a8b02)