[INTERNAL] transformApiJson: Fix reading docuindex.json
The relative path within a .library file that points to the
docuindex.json uses runtime paths, including the "test-resources"
segment.
When running a JSDoc build on the library sources, this path can't be
resolved. Therefore a rewrite of the path is done, to point to the file
with its actual filesystem path.

JIRA: CPOUI5FOUNDATION-518
Change-Id: Ib1829484789171502e48c153ed7184eddbdf5b0c