[FIX] sap.f.DynamicPage: Issue with header state corrected
Problem:
When the header of the DynamicPage is hidden
with property "visible" set to "false")
and we scroll down the content
=> the title is snapped (es expacted).
However, when we scroll back to the top,
the title is not expanded back.
Further, if we show back the header at this stage
(with property "visible" set to "true")
=> we end up with snapped title and expanded header
at the same time (inconsistent state)

Solution:
Ensured that we expand the page back upon scrolling to top
(symmetrical to the snapping of the page upon scrolling down).
Note:
We toggle the "expanded" state of the page upon scroll
even when header is hidden because of the other top items
(title and sticky subheader) that still make sense to
snap/expand upon scroll even when no visible header.

Change-Id: I112051cf540c386f8a2912980a102331e20bc3bf
BCP: 2280133870