[FIX] sap.uxap.ObjectPageLayout: Prevent error when footer is set
Problem: When "footer" is set and ObjectPageLayout is invalidated at the
same time, an error is thrown, as the footer is still not rendered, but
in onBeforeRendering we check the footer height.

Solution: Check for DomRef available, before getting its height. The
footer height will still be got into account after it is rendered.

BCP: 2270096171
Change-Id: I0577709db6777c8b5e69994f0a5ecf4fb5274f70