[INTERNAL] MDCTable: Invalidate inner table after applying column changes
Invalidation is suppressed in MDC when applying aggregation changes.
As a result of this behavior, #invalidate will not be called on the
inner table, which can lead to problems in cases where the inner table
expects this call.
It is now called manually after all changes have been applied.

Change-Id: Ia7ce91daa72b4671560104aa76b06970c3db4136
BCP: 2270051807