[INTERNAL] sap.m.IllustrationPool: Stabilized test on FireFox browser
Problem: The sap.m.IllustrationPool qUnit test isn't passing on the
latest version of the FireFox browser.

Root Cause: In the IllustrationPool we are setting the aria hidden
attribute with the "ariaHidden" API which isn't supported in FF.

Solution: We use API which is supported across all browsers
(setAttribute(name, value)).

BCP: 2280137915
Change-Id: Idf3993fedca1d832de5562034068e751c6e57fab