[INTERNAL][FIX]BaseTreeModifier: Remove error message
Some controls are generated without stable IDs inside templates and are
not part of the template (e.g. sap.fe.core.controls.FieldWrapper).

These cases were raising an error on the console during the control ID
check in templates, so we removed it as preventing the adaptation
(returning "false" on the check) is enough.

Change-Id: Idd50291d2237ba7c0a6b0df207a9b8abb80a97a6
BCP: 2280097003