[INTERNAL][FIX] CellSelector: Add EPSILON value to unit tests
When running the unit tests for the CellSelector plugin, some failed for Firefox. This was due to getBoundingClientRect() returning different width values than the width in the defined CSS style.
This issue is due to the browser.
To prevent the unit tests from failing, an EPSILON value of 1px was introduced, to allow a difference of 1px between actual canvas width and expected width.

Change-Id: Id54a39fb272158ae39325fc763322e3f6d803596