[INTERNAL][FIX]sap.ui.fl, rta: Compatibility Fixes
Some consumers (Visual Editor) are still calling "getDefinition" on
objects which were already converted to FlexObject (e.g. Variants), as
they all used to be "changes" before (and have the same structure).
To avoid errors we are temporarily adding the method to the FlexObject
until all consumers are adjusted.
Note: contrary to the previous implementation, this returns a new
object, meaning that mutations on the return object won't be considered.

Also the Stack initialization expected the changes to have a selector,
which is not the case for Variants.

Change-Id: Idd34ea740f3bed82d6e1203dc70a2f8b3a8835b1
BCP: 2280139410
CR-Id: 002075125900002454012022