[INTERNAL] FilterController: use delta appliance for incoming state
In case of multi valued fields, the StateUtil & FilterController handled
the values of a FilterField as snapshot.

The rest of StateUtil related functionality (such as sorting, grouping,
items, etc.) is handled as additive delta.

This functionality of StateUtil#applyExternalState has now been streamlined
with the filter part in the state object. When a filter should be removed,
the exact condition needs to be explicitly marked as filtered: false

FilterableListContent is now using StateUtil#diffState to merge
the initial state from the delegate into the current state once
the VH gets openened. This ensures that only the exact delta is
going to be applied by StateUtil.

Jira: CPOUIFDENVER-1803
Change-Id: I35f61f82012779f37091bb82b48a73510d9e3347