[INTERNAL] mdx.flexibility: several improvements
- move the ItemBaseFlex#_delayInvalidate functionality to the central
flexibility/Util file to always start the suppression of invalidation
on start of an incoming chain of change appliance --> Once all pending
changes have been applied the invalidation will be triggered manually by
the changehandler.

- ConditionFlex no longer takes care of inserting the filterCondition
changes into the ConditionModel, this part is now handled by the FilterBar.
Once the filterConditions property has been updated, the FilterBar
will trigger a sync in FilterBar#_onModifications

- ItemBaseFlex#_applyRemove does now use "KeepDom" when destroying an item.
For example if a large amount of columns is being destroyed, this will take
care to only update the DOM after the next manual invalidation triggerd by
the changehandler instead of the control itself.

BCP: 2270087432
Jira: CPOUIFDENVER-1792

Change-Id: Ib69a0b50f200e2013f008970c270124bc3c3e3ea