[INTERNAL] FilterBarBase: stabilize timings
Fixes failing FE tests. Since the latest changes in flex orchestration
in FBB the marker bIgnoreChanges is only used for the appliance of
initial filters. Resetting this flag in the according promise ensures
the proper orchestration.

The PersistenceProvider in addition does not properly propagate the
VariantModel to the VM through the DefaultProviderRegistry, this
needs to be adressed in a follow-up change in DPR/PersistenceProvider.
For now the PersistenceProvider has been set to visible: false again,
to avoid false VM references.

Change-Id: I3e4c2994fb5768bb460b78915cd186f8221297eb