[FIX] sap.ui.core.Configuration: Disable timezone configuration
Deactivate the possibility for central timezone configuration since it
causes compatibility issues.

Attempts to changing the timezone configuration are ignored. For the
time being, the browser timezone is always used.

Note that the possibility to pass a custom timezone to
sap.ui.core.format.DateFormat.DateTimeWithTimezone instances is unaffected by this.

JIRA: MI2C-3095
Change-Id: I731ac6d46b7583c8e41ddf4195209716938709fe