[INTERNAL] sap.m.Table: ACC - Announce the required state of Label in Column
When the column header contains a Label, which has the property "required" set to true, the announcement "Contains required content" is now made.

sap.m.Label#getAccessibilityInfo was extended to include a required property, which can either be true or false depending on the required state of the label control.

JIRA: CPOUIFTEAMB-1512
Change-Id: I20bf8994e633d534c4988777dae04ba6a1c21a6f