[INTERNAL] Demo Kit: SearchField blur on mobile fixed
Problem: The soft keyboard was closing for a split second when the search dialog was initially open. This was caused by focus switching between the main and in-dialog SearchFields.

Solution: Prevent focus from being applied on the main SearchField when a non-desktop device is used, and instead, directly open the search dialog.

BCP: 2280129599
Change-Id: I191490bb734546025234e0bda72fd7adb53fdac2