[FIX] sap.ui.core.ScrollEnablement: Scroll-end callback issue corrected
Problem:
It is possible to arrive at a case where the callback
for end of scroll (fnScrollEndCallback) is never called.
This happens when the scroll stops before its completion
(e.g. due to concurrent DOM modifications) and then
the underlying jQuery.animate does not fire the "complete"
callback, but does fire the "fail" and "always" callbacks
if subscribed.

Solution:
Subscribe to the "always" callback of jQuery.animate
to always get notified what the scroll ends either
completely or stopped early (due to concurrent DOM modifications).

BCP: 002075129500004009642022
Change-Id: I9b0b1df1b7e4f963bbc003101228be11b50ed433