[FIX] sap.m.OverflowToolbar: OverflowToolbarButton invalidation prevented
Problem: OverflowToolbarButton is invalidated onBeforeRendering, which
causes an error, when trying to open the OverflowToolbar's Popover menu.
This leads to resetting of the OverflowToolbar and rerender. As a
result, the Popover is getting closed (not correct behavior), but
unsuccessfully, as its opener is rerendered.

Solution: The problem happens because of a new implementation in the
Button control, which sets "accessKey" property in the onBeforeRendering
logic, which invalidates the control.
"accessKey" property should be added in the noInvalidationProps of
OverflowToolbarButton's config, so the OverflowToolbar won't be
invalidated, due to its change, as it does not affect size.

BCP: 2280160553
Change-Id: I07b58f25d080f5faf6a80a29b0509682cf19e0ce