[INTERNAL][FIX]sap.ui.rta: CViz AppDescriptor Fix
The Change Visualization was not filtering out AppDescriptor changes,
which don't have selectors and therefore were breaking the application.

In this change also a small issue with the icon tooltip was fixed, where
the command name was being used instead of the category name when
retrieving the category text.

Change-Id: I90ba13fe81a88359c4175bbbf347de111cfd3725
BCP: 2280164441