[INTERNAL] sap.ui.rta: stack fires modified event only once during pushAndExecute
the fireModified call was moved from the generic push function to the pushExecutedCommand function.
With this in both scenarios (pushExecutedCommand and pushAndExecute) the event is only fired once.
Calling the push function directly will not fire the event anymore.
Calling the remove function directly will also not fire the event anymore (pop should be used).

BCP: 2280160445
Change-Id: Ibee30efabbf6edbf4047fe5b7343ef23708cf6cc