[INTERNAL] sap.ui.rta: tests adaptation
Problem:
The SmartLabel gets initialised on SmartField initialise event. If the
SmartField has two nested base controls the initialise event will
be triggered with delay after the second base control gets created. This
makes possible that SmartField has inner control but the SmartLabel
still waits for the initialised event to be fired.

Solution:
We check not only for the first inner control. We also check if this
control has assigned Label.

Change-Id: I805f09bc5cbc3f82665625f019355df620aa963a