[FIX] mvc.Controller: Cleanup dangling fragment content before exit
Fixes a bug in the controller that prevented fragments to be cleaned up.
Previously the automatic async cleanup of the managed fragments created with
loadFragment() API was coupled to the existance of an onExit hook.

The correct behavior is to always try to clean up if possible, whether a
hook is implemented or not.

The change also includes stabelizing of the unit tests and
additional checks if all dangling controls have been destroyed correctly.

Change-Id: I3f4e617daf6520cd9fa44ec8eb0ce73c662177ca