[FIX] sap.m.TabContainer: prevent itemSelect event
Issue:
On 'change' event there are a 'sap.ui.base.Event'
instances are received inside the handler. The framework
events do not have an 'isDefaultPrevented' method.

Solution:
We are expecting a native browser events in this case,
which are wrapped with jQuery and do support the
jQuery.Event.prototype.isDefaultPrevented method.

BCP: 1980097573
Change-Id: Ibf46d75b2b74a7b7ab80091c941894d6fab0465c