[FIX] sap.ui.model.odata.ODateMetaModel#fetchCodeList: don't reuse codelist promise over models if it is not yet resolved
If an ODataModel is created and the code list request is triggered but
not yet resolved. Before the response returns the model gets destroyed.
In that case the cached code list promise must not be used by other
models for the same service because that promise never resolves.

PS1: Downport to 1.102

Change-Id: Id8a6b08754b0c4cc37e975c246eb9b294ff73549
BCP: 2280171731
CR-Id: 002075125800003400682022
(cherry picked from commit c2bebfe38a7110d648499c19610d77feabb4415e)