[INTERNAL][FIX] sap.uxap.ObjectPageSection: Fixed 'showMore' alignment
Problem: Visual tests report that 'showMore' button is not aligned as
expected on Quartz Light theme.

Root Cause: There is a small typo caused by a regression (5615944).

Solution: We fix the regression (missing comma in the ObjectPageSection
Quartz Light CSS).

Additionally we apply the now fixed styles for the new heading
aggregation (the feature which originally caused the regression) that we
have in Quartz Light to Quartz Dark too for consistency.

Change-Id: Ia8b2ad09a28c71c212fad2c9a68cead744e4fa5c