[INTERNAL] sap.f.semantic.SemanticPage: stabilize QUnit tests
Problem: Some tests are failing on nightly built.

Root cause: An assertion, based on a keyboard event fired with the `ctrlKey` property, was failing on Safari because `metaKey` is the analog there. As a result, the event wasn't triggering the expected keyboard shortcut.

Solution: Make the keyboard event compatible with macOS.

JIRA: BGSOFUIPIRIN-5845
BCP: 2280160890
Change-Id: I354562910c54c2a5138c5b6a7b81fcaba6cb50b8