[INTERNAL][FIX] sap.m.Select: Fixed incorrect color of icon
Problem: On Morning/Evening Horizon, when we have 'iconOnly' Select
in sap.m.Bar the color of the icon doesn't match the icon color of the
sap.m.Button when it's also in icon only mode.

Solution: For 'iconOnly' Select we apply the same icon color which is
applied for icon only Button.

BCP: 2280148589
Change-Id: I301842aaf4047d3a8102cb8d2ca6438c501353d4