[FIX] sap.m.Select: Fixed incorrect typing behavior
Problem: If the user starts typing and selects an item, but continue
to input characters which are still valid string starting sequence for
the newly selected (now current) item, the selection is moved to the
next item.

Solution: We make sure to keep the selected item if the user input keeps
being valid for the current selection within the type timeout (1sec).

BCP: 2280171879
Fixes: https://github.com/SAP/openui5/issues/2713
Change-Id: Ie4a2d23aa7a0e49152374e9185f1fddd1ef36676