[FIX] Field: error after entering number
If a Field is bound to a Unit or Currency type 2 input controls are rendered, one for number, one for unit.
If the Field is initially empty and the user enters the number an error occurs as the unit is still missing.
If then a unit is entered the number gets cleared.

Change-Id: Ib1c2f848fb12faa551d849c6407efc89bd6c30d4
BCP: 2280171152