[INTERNAL] Table: hidden column header row should not be focusable
When columns in the table do not have "header" aggregation defined, then the table hides the column header row.
In such a case, the column header row should not be focusable, else, there is a side effect that the focus is lost.
For example, when the user presses the "END" key on the keyboard, focus jumps to the last visible column and upon pressing "HOME" the focus is set on the hidden column header row which is unexpected.
Now, the Table correctly sets the itemDomRefs only for the focusable and visible elements of the table.
The table will now set the focus on the "first" visible item, if the table header row is hidden for the above explained example.

Change-Id: I006201d9005871cec6d4d5f97bf933fe49a770c5