[INTERNAL] TreeTable: Skip unstable test
Sometimes, the MockServer response comes too early, and instead of 2
rowsUpdated events that are expected by the test, only 1 is fired -
trying to fire multiple rowsUpdated events within a timeout of 0ms is
condensed to one for performance reasons.

Increasing the response time of the MockServer breaks other tests.

The unstable test is skipped until a solution is found.

Change-Id: I827144a8fd6c8c66bf1eab2725efbcdbdd4cd2cb
BCP: 2270119656
CR-Id: 002075125800003761242022