[FEATURE] SimpleForm: stable IDs
Use stable IDs for FormContainers and FormElements.

The ID of the FormContainer is based on the ID of its Title or ToolBar.
If no Title or Toolbar it gets a static ID.
(Only the first FormContainer can exist without Title or Toolbar.)

The ID of the FormElement is based on the ID of its Label.
If there is no Label it gets a static ID.
(Only the first FormElement of a FormContainer can exist without a Label.)

Fixes an error with the destruction of overlays in sap.ui.dt.DesignTime.

Change-Id: Ibadd64d290303a435e99de7f236333c067b01677
JIRA: CPOUIFDENVER-1797