[INTERNAL] FilterBarBase: refactor _setXConditions to apply a delta
When applying changes to the FilterBarBase, it will no longer clear
the complete conditions in the CM but only calculate the exact delta
between the state before and after the modification.

This delta is then fine granularly applied using
ConditionModel#addCondition and
ConditionModel#removeCondition

This behaviour will improve the handling to not modify or reapply
unchanged conditions.

BCP: 2270112782
Change-Id: Idd736d308a51c0d2228aae32eea4f83fa84a5e55