[FIX] sap.uxap.ObjectPageLayout: fire "navigate" event later
Problem: When user clicks on an AnchorBar tab, and the application wants
to redirect to another tab (in some circumstances), using the "navigate"
event and setting a new selected Section in this handler, the selected
Section is correct, but the AnchorBar selected button is wrong.

Solution: Fire the "navigate" event, after the call of
requestAdjustLayout, so that if an app wants to programatically change
the selected Section in their "navigate" event handler, the using the
setSelectedSection API, the selected AnchorBar button is correct.

BCP: 2280115969
Change-Id: I56177acb80dc72d19de75c9ee1f785d57ca5aab7