[FEATURE] sap.ui.core.format.DateFormat: introduce calendarWeekNumbering format option
Introduce the calendarWeekNumbering format option for simplicity. It overrules
the existing format options firstDayOfWeek and minimalDaysInFirstWeek.

Its values are defined in enum sap.ui.core.date.CalendarWeekNumbering:
- Default
    The framework determines the week numbering scheme; currently
    it is derived from the active format locale.
- ISO_8601
    Official calendar week numbering in most of Europe (ISO 8601
    standard)
- WesternTraditional
    Official calendar week numbering in the United States, Canada,
    Brazil, Israel, Japan, and other countries
- MiddleEastern
    Official calendar week numbering in much of the Middle East

A utility class with a function is introduced which can resolve the enum values.
This is useful for controls which need to know the first day of the week for when
they only get the enum as parameter.

Examples
* DateFormat:
DateFormat.getDateInstance({
	pattern: "Y'-'w",
	calendarWeekNumbering: CalendarWeekNumbering.ISO_8601
});

* CalendarUtils:
CalendarUtils.getWeekConfigurationValues(CalendarWeekNumbering.ISO_8601, deLocale);
// {firstDayOfWeek: 1, minimalDaysInFirstWeek: 4}

Can be tested with:
http://localhost:8080/test-resources/sap/ui/core/DateFormat.html

using
{
	pattern: "Y'-'w",
	calendarWeekNumbering: CalendarWeekNumbering.ISO_8601
}

JIRA: CPOUI5MODELS-1018
Change-Id: I40bf9013ef8e76b7d934b89d893d8ae4153da5e9