[FIX] util/fetch: add protocol back to protocol-relative URL string
Cloud portal interceptor can't handle protocol-relative URLs (the ones that start directly with "//" without having a preceeding protocol). The missing protocol was added within jQuery.ajax implementation and we have to do the same in fetch.js.

The fix needs to be downported to rel-1.105 although the attached incident doesn't have priority high or above. This is because all URLs given to XMLHttpRequest that start directly with "//" can't be handled correctly when UI5 runs within Cloud Portal. This could affect applications in different aspects.

BCP: 002075129500005263542022
Change-Id: I9936183ef9701fb13327384a2533f951b5e8e996
CR-Id: 002075125800003674212022