[INTERNAL]sap.ui.fl: Collect 'delete' calls
When many "delete" changes are done at once (e.g. with a local reset),
we currently call the "delete" endpoint multiple times instead of
using the "condense" route that allows multiple deletes to be sent at
once.

This change collects the deletes and uses the "condense" endpoint to
decrease the number of calls to the backend.

Change-Id: I6d8d5228bb2a1d11de3215cfbef3e77beb134e45
JIRA: TEAM42-226