[FIX] Demo Kit: View subclasses link is fixed
In some rare cases, the Documentation router global click handler
id triggered prior to specific link handler and the "prevent default"
option is not set accordingly. If we add the option to the click event
of the link it should be 'true' by the time the click reaches global
handler.

BCP: 2280170250

Change-Id: Ie0933947c7237ea8dedc245103a4ceb3ca331df3