[FIX] sap.uxap.ObjectPageSectionBase: Correct lifecycle issue
Background:
The ObjectPageSectionBase has a _getGrid method
that lazy loads the inner grid upon need.

Problem:
Calls to the _getGrid method on *destroyed* (sub)sections
still leads to the creation of a new instance of the grid.

Solution:
Added a check for the state of the section in order to avoid
the creation of the grid if the (sub)section is being destroyed

Alternative solutions (to split the creation of the grid
from the _getGrid getter function) were also tried but
discarded, e.g:
1) moving the creation into the "onBeforeRendering" hook is
not a solution as the _getGrid getter may need to be called
earlier than the onBeforeRendering hook [e.g. by the
aggragationForwarder]
2) moving the creation into the "init" hook was also
discarded since it changes the lifecycle (causes the grid
to be rendered earlier than before) that may cause regressions
in existing apps.
This is why the current solution (the check against the flag,
as already used in many controls) was chosen.

BCP: 2270125667
Change-Id: I1e37c27441827d282fc480a8653a1cf2a1d88828