[INTERNAL][FIX]SimpleForm Headerless Group Move Fix
When moving a Headerless group in SimpleForm, the change was not applied
on reload because an invalid selector was being saved as
dependentSelector (id was not found for non-existing header).

Since a title element is always created by the change handler in this
case, the dependentSelector can be removed from the change content.

Change-Id: I506802e470f7e609e278ed7657f10749e91c48a4
BCP: 2270119839