[INTERNAL][FIX]sap.ui.dt: Omit DT error during Overlay creation
In cases where many elements are removed and added in a row (e.g. move
in SimpleForm), an error was being displayed on the console when the
parent overlay was already destroyed during the creation of its children
overlays. As this is a normal situation with no side effects, the
related error message is not logged anymore.

Change-Id: I5b200817d1a6c17653b075e514ab07e2f3b79844
BCP: 2270131664