[INTERNAL] sap.ui.mdc: temp. fix FilterBar opa test
The test module qunit/filterbar/sample/opa/opaTests.qunit.ja uses
sap.ui.require instead of sap.ui.define, but does not prevent a QUnit
autostart. This causes a race condition and sometimes the test starts
and finishes with a still empty queue.

As a temp. workaround, autostart is now set to false in the testsuite.
But either the test should be fully migrated to the test starter (not
have an own, full fledged HTML page) or use the HTML page, not a module.

Change-Id: I8d5b7392bd2be25a800deccfddaeb5656d1ac83b