[INTERNAL] Enhance ACC handling for required state
* Introduce a required flag in the ACCInfo which can be used as alternative
to a verbal announcement in the description.
* Label: Use this flag in ACCInfo and let consumers decide on the required
announcement
* Label: Provide function setIsInColumnHeaderContext to inform the Label
to be in an ColumnHeader context (ensures aria-hidden is used on : and * like
in editable form context)
* Enable consumers of the ACCInfo (Tables, List) to announce the required
state
* If the table column header contains a sap.m.Label control with the
required property set to true then the "Contains required content"
announcement must be provided only when the focus is put on the header
row or directly on the active column header. Otherwise a standard required
announcement should be done.
* InputBase: Use translatable text for required announcement.
* DateTimePicker, TimePicker: Avoid access of aria-label attribute of the label

BCP: 2270132228
JIRA: CPOUIFTEAMB-1512

Change-Id: If9d7000fb2624b34b210b5e96a7dc54a16310254