[INTERNAL][FIX] sap.m: Destroyed SemanticPage should not recreate page
Background:
The SemantciPage has a _getPage method
that lazy loads the inner page upon need.

Problem:
Calls to the _getPage method on *destroyed* SemantciPage
still leads to the creation of a new instance of the inner page.

Solution:
Added a check for the state of SemantciPage in order to avoid
the creation of the inner page if the SemantciPage is being destroyed

Alternative solutions (to split the creation of the page
from the _getPage getter function) were also tried but
discarded, e.g:
1) moving the creation into the "onBeforeRendering" hook is
not a solution as the _getPage getter may need to be called
earlier than the onBeforeRendering hook [e.g. by the
aggragationForwarder]
2) moving the creation into the "init" hook was also
discarded since it changes the lifecycle (causes the inner page
to be rendered earlier than before) that may cause regressions
in existing apps.
This is why the current solution (the check against the flag,
as already used in many controls) was chosen.

CR-Id: 002075125800003973032022
Change-Id: Idd0ee9c2aa7beec7ae360ed69ac82c81457d987a
BCP: 002075129500006787062022
BCP: 2270135825
(cherry picked from commit 833fa8ee9af3f8efa22019a640f20387bddcffa5)