[FIX] Field: Change-Promise pending if validation error on unit
If a Field is bound to a Unit or Currency type two Input controls are rendered, one for number and one for unit/currency.
If there is a value help assigned to the unit or currency this is used to parse and validate the entered value. This parsing might be asynchonous. If in this asynchrounus parsing then leads to a validation error the Promise returned by the change-event is pending forever. The rejection is not triggered.

Change-Id: I0cd2a8ad22cb41da406b127fc7a1d56cd5d3498f
BCP: 2270123293